detection_pipeline/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
albumentations==1.3.0
alembic==1.10.2
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.2.0
backcall==0.2.0
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
configobj==5.0.8
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
databricks-cli==0.17.6
debugpy==1.6.6
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
docker==6.0.1
dpath==2.1.5
dulwich==0.21.3
dvc==2.52.0
dvc-data==0.46.0
dvc-gdrive==2.19.1
dvc-http==2.30.2
dvc-objects==0.21.1
dvc-render==0.3.1
dvc-studio-client==0.6.1
dvc-task==0.2.0
entrypoints==0.4
executing==1.2.0
filelock==3.10.0
Flask==2.2.3
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
funcy==2.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.59.0
grandalf==0.8
greenlet==2.0.2
grpcio==1.51.3
gunicorn==20.1.0
httplib2==0.22.0
hydra-core==1.3.2
idna==3.4
imageio==2.27.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.3.0
kiwisolver==1.4.4
kombu==5.2.4
lazy_loader==0.2
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==2.2.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nanotime==0.5.2
nest-asyncio==1.5.6
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
orjson==3.8.9
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pillow-heif==0.10.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyDrive2==1.15.2
pygit2==1.12.0
Pygments==2.14.0
pygtrie==2.5.0
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyre-extensions==0.0.30
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.1
qudida==0.0.4
querystring-parser==1.2.4
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
scmrepo==0.2.0
seaborn==0.12.2
shap==0.41.0
shortuuid==1.0.11
shtab==1.6.0
six==1.16.0
scikit-learn==1.2.2
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==2.0.8
sqlparse==0.4.3
sqltrie==0.3.0
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.21
tomlkit==0.11.7
torch==2.0.0
torcheval==0.0.6
torchtnt==0.0.7
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.2.3
yarl==1.8.2
zc.lockfile==3.0.post1
zipp==3.15.0
